requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
Pillow>=9.0.0
matplotlib>=3.5.0
transformers==4.39.0
opencv-python>=4.6.0
tqdm>=4.66.0
hydra-core>=1.3.2
iopath>=0.1.10 